Add config option for hiding password in preview
By default all passwords read during the fuzzy search will be displayed
in fzf's preview pane.

In some situations, this can be undesirable. For example if you are
fetching a password while recording or having bystanding people watch your
screen.

To alleviate this, an additional configuration option is added, which,
when enabled, will hide the password from the preview pane and only show
the additional information stored in the password entry (e.g. login,
url, etc.).

Currently the setting defaults to `off`, so the default behaviour is
unchanged.

However it might be worth considering to have it default to `on`,
providing more privacy by default.