Add support for hiding preview pane
Password-Store paths can get quite long and depending on your tmux pane
layout, the tmux-pass split might be quite small, which makes it harder
to identify the correct path.

To help with this, a keybinding is added to the fzf call (`TAB`) which
allows the user to toggle preview pane visibility.

Additionaly an additional tmux configuration option `@pass-hide-preview`
allows users to configure the default state of the preview pane.

The setting defaults to `off` (i.e. the preview pane is shown), thus
keeping the default behaviour unchanged.
If set to `on`, the preview pane will be hidden by default.