Add support for copying username
The data organization scheme favored by the password-store author, suggests
putting the password on the first and additional information on subsequent
lines of a password-store entry.

These additional fields are expected to be one key / value pair per line,
separated by a colon `:` (possibly with a space after the colon).

For those, using this scheme, it would be convenient to being able to not only
copy the password, but also the username reuqired for logging in.

To achieve this, an additional key mapping (`alt-enter`) is added, which will
copy the username instead of the password.

Since the key used for username fields may vary (e.g. *login*, *user*,
etc.), a list of keys is tried and the first (case insensitive) match is taken.